requirements.txt
click==8.2.0
colorama==0.4.6
Flask==2.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.31
protobuf==3.20.1
psycopg2==2.9.10
python-dotenv==0.21.0
Werkzeug==2.2.3
